# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: bloq5 — Plateforme de gestion immobilière
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rental-requests:
    get:
      operationId: listRentalRequests
      tags: [rental-requests]
      summary: Get current user rental requests
      responses:
        "200":
          description: Rental requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RentalRequest"

    post:
      operationId: createRentalRequest
      tags: [rental-requests]
      summary: Create a rental request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRentalRequestBody"
      responses:
        "201":
          description: Created rental request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalRequest"
        "400":
          description: Bad request (max 3 requests or payment required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rental-requests/{id}:
    get:
      operationId: getRentalRequest
      tags: [rental-requests]
      summary: Get a specific rental request detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rental request detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalRequestDetail"
  /rental-requests/{id}/status:
    patch:
      operationId: updateRentalRequestStatus
      tags: [rental-requests]
      summary: Update rental request status (owner/manager)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRentalRequestStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalRequest"

  # Requests by property (for owner/manager)
  /properties/{id}/rental-requests:
    get:
      operationId: listPropertyRentalRequests
      tags: [rental-requests]
      summary: List rental requests for a property (owner/manager)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rental requests for property
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RentalRequest"

  # --- FAVORITES ---
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    RentalRequest:
      type: object
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        propertyTitle:
          type: string
        propertyImage:
          type: string
        userId:
          type: string
        status:
          type: string
          enum: [pending, in_review, awaiting_documents, approved, rejected]
        message:
          type: string
        applicantName:
          type: string
        applicantEmail:
          type: string
        applicantPhone:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - propertyId
        - propertyTitle
        - propertyImage
        - userId
        - status
        - message
        - applicantName
        - applicantEmail
        - createdAt
        - updatedAt
    RentalRequestDetail:
      allOf:
        - $ref: "#/components/schemas/RentalRequest"
        - type: object
          properties:
            conversationEnded:
              type: boolean
            hasUnreadMessages:
              type: boolean
          required: [conversationEnded, hasUnreadMessages]
    CreateRentalRequestBody:
      type: object
      properties:
        propertyId:
          type: integer
        message:
          type: string
        applicantName:
          type: string
        applicantEmail:
          type: string
        applicantPhone:
          type: ["string", "null"]
      required: [propertyId, message, applicantName, applicantEmail]
    UpdateRentalRequestStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_review, awaiting_documents, approved, rejected]
        note:
          type: ["string", "null"]
      required: [status]
...