# lib/api-zod/src/generated/types/rentalRequestDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * bloq5 — Plateforme de gestion immobilière
 * OpenAPI spec version: 0.1.0
 */
import type { RentalRequest } from "./rentalRequest";

export type RentalRequestDetail = RentalRequest & {
  conversationEnded: boolean;
  hasUnreadMessages: boolean;
};