# lib/api-zod/src/generated/types/updateRentalRequestStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * bloq5 — Plateforme de gestion immobilière
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateRentalRequestStatusBodyStatus } from "./updateRentalRequestStatusBodyStatus";

export interface UpdateRentalRequestStatusBody {
  status: UpdateRentalRequestStatusBodyStatus;
  /** @nullable */
  note?: string | null;
}